--- client/src/App.css
@@ -1,15 +1,0 @@
-@import url(https://fonts.googleapis.com/css?family=Exo:100,200,400);
-@import url(https://fonts.googleapis.com/css?family=Source+Sans+Pro:700,400,300);
-
-.bg_color {
-  background-color: rgb(70, 70, 70);
-}
-
-.header_size{
-  font-weight: bold;
-  font-family: 'Exo', sans-serif;
-}
-
-.insta_color{
-	color: #b47c22 !important;
-}
