+++ client/src/App.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {BrowserRouter as Router, Route} from 'react-router-dom';
-import './App.css';
+import './App.scss';
 import Navbar from './components/layout/Navbar';
 import Landing from './components/layout/Landing';
 import Footer from './components/layout/Footer';
@@ -10,12 +10,12 @@ import Register from './components/auth/Register';
 function App() {
   return (
     <Router>
+      <Navbar/>
     <div className="App">
-      <Navbar />
       <Route exact path= "/" component={Landing} /> 
       <Route exact path="/register" component={Register} />
-      <Footer />
     </div>
+    <Footer/>
     </Router>
   );
 }
