+++ client/src/components/auth/Register.jsx
@@ -1,0 +1,123 @@
+import React, { Component } from "react";
+import axios from "axios";
+import classnames from "classnames";
+import loginImg from "../../register1.svg";
+
+export class Register extends Component {
+  constructor() {
+    super();
+    this.state = {
+      name: "",
+      email: "",
+      password: "",
+      password2: "",
+      errors: {}
+    };
+
+    this.onChange = this.onChange.bind(this);
+    this.onSubmit = this.onSubmit.bind(this);
+  }
+
+  onChange(e) {
+    this.setState({ [e.target.name]: e.target.value });
+  }
+
+  onSubmit(e) {
+    e.preventDefault();
+    const newUser = {
+      name: this.state.name,
+      email: this.state.email,
+      password: this.state.password,
+      password2: this.state.password2
+    };
+
+    axios
+      .post("/api/users/register", newUser)
+      .then(res => console.log(res.data))
+      .catch(err => this.setState({ errors: err.response.data }));
+  }
+
+  render() {
+    const { errors } = this.state;
+
+    return (
+      <div className="base-container" ref={this.props.containerRef}>
+      <div className="header">Register</div>
+      <div className="content">
+        <div className="image">
+          <img src={loginImg} />
+        </div>
+        <form onSubmit={this.onSubmit}>
+          <div className="form">
+            <div className="form-group">
+              <input
+                    type="text"
+                    className={classnames("sign_up_text", {
+                      "is-invalid": errors.name
+                    })}
+                    placeholder="Name"
+                    name="name"
+                    value={this.state.name}
+                    onChange={this.onChange}
+                  />
+                  {errors.name && (
+                    <div className="invalid-feedback">{errors.name}</div>
+                  )}
+              </div>
+            <div className="form-group">
+              <input
+                    type="email"
+                    className={classnames("sign_up_email", {
+                      "is-invalid": errors.email
+                    })}
+                    placeholder="Email Address"
+                    name="email"
+                    value={this.state.email}
+                    onChange={this.onChange}
+                  />
+                  {errors.email && (
+                    <div className="invalid-feedback">{errors.email}</div>
+                  )}
+             </div>
+            <div className="form-group">
+                <input
+                  type="password"
+                  className={classnames("form-control form-control-lg", {
+                    "is-invalid": errors.password
+                  })}
+                  placeholder="Password"
+                  name="password"
+                  value={this.state.password}
+                  onChange={this.onChange}
+                />
+                {errors.password && (
+                  <div className="invalid-feedback">{errors.password}</div>
+                )}
+             </div>
+             <div className="form-group">
+              <input
+                  type="password"
+                  className={classnames("form-control form-control-lg", {
+                    "is-invalid": errors.password2
+                  })}
+                  placeholder="Confirm Password"
+                  name="password2"
+                  value={this.state.password2}
+                  onChange={this.onChange}
+                />
+                {errors.password2 && (
+                  <div className="invalid-feedback">{errors.password2}</div>
+                )}
+             </div>
+             <input type="submit" className="btn btn-info btn-block mt-4" />
+          </div>
+          </form>
+        </div>
+        
+      </div>
+      );
+    }
+  }
+  
+  export default Register;
+
