+++ client/src/components/auth/index.jsx
@@ -1,0 +1,4 @@
+import "./style.scss";
+// import '../../App.scss';
+export { Login } from "./Login";
+export { Register } from "./Register";
